+++ changelog.md
@@ -11,6 +11,9 @@
   - Defaults to an indent of 4 that starts from the beginning of the line or
     after a block quote start
   - Updated documentation for Md007 to mention Pml101 and when to use it
+- [Issue 983](https://github.com/jackdewinter/pymarkdown/issues/983)
+  - Added base foundation for new documentation, publishing on
+    read-the-docs.
 
 ### Fixed
 
@@ -30,6 +33,10 @@
 - [Issue 981](https://github.com/jackdewinter/pymarkdown/issues/981)
   - added documentation for Rule Pml100
   - cleaned up mentions of whitespace in pragma and front matter extensions
+- [Issue 986](https://github.com/jackdewinter/pymarkdown/issues/986)
+  - initial setting of log level to DEBUG with `--stack-trace` command line
+    flag not working as expected
+  - fixed to properly set log level to DEBUG for early application debugging
 
 ### Changed